shell script to feed a list of videos downloader
youtube-comment-downloader is used to download the comments of a video. start.sh parses a list of videos, iterates on the list and downloads the comments, then using jq converts the json to tsv. 

TSV being used as that is the best format to upload these files into google sheets for MonkeyLearn to process.